modularise the code in the event route
