+++ src/routes/event.js
@@ -1,13 +1,5 @@
-const { mockEvents } = require('../../database/events_schema.js');
+const findEvent = require('../actions/find_event.js');
 
 module.exports = (req, res) => {
-  const id = req.params.id;
-  mockEvents.findOne({_id:id}, (error, event)=> {
-    if(error) res.render('not_found');
-    else {
-      res.render('event', {
-          event: event
-      })
-    }
-  })
+  findEvent(req.params.id, 'event', res);
 }
